requirements.txt
asgiref==3.10.0
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
